+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change password creation to be disabled by default in server creation
+- Always create a password when password delivery method is chosen
 
 ## [1.1.0] - 2021-06-03
 ### Added